{{cookiecutter.project_slug}}/tests/{{cookiecutter.project_slug}}/test_greet.py
import pytest

from {{cookiecutter.project_slug}} import greet


@pytest.mark.parametrize(
    ("name", "expected"), [("John", "Hello, John!"), ("123", "Hello, 123!")]
)
def test_greet(name: str, expected: str) -> None:
    """Test the greet function."""
    assert greet(name) == expected
