{{cookiecutter.project_slug}}/pyproject.toml
[project]
name = "{{cookiecutter.project_name}}"
authors = [{name = "{{cookiecutter.author_name}}", email = "{{cookiecutter.author_email}}"}]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
# those are defined in the top level __init__.py and it's docstring
dynamic = ["version", "description"]
# https://iscinumpy.dev/post/bound-version-constraints/#pinning-the-python-version-is-special
requires-python = ">={{cookiecutter.minimum_python_version}}"
dependencies = []

# [[tool.uv.index]]
# name = "zalando"
# url = "https://pypi.zalando.net/repository/pypi/simple"
# default = true

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pygments>=2.19.1",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["flit_core >=3.12,<4"]
build-backend = "flit_core.buildapi"

[tool.ruff]
line-length = 88
target-version = "py{{cookiecutter.minimum_python_version.replace('.', '')}}"
exclude = []
[tool.ruff.format]
preview = true
[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = [
    "CPY",
 # https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "COM812",
    "COM819",
    "D206",
    "D300",
    "E111",
    "E114",
    "E117",
    "E501",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "W191",
]
[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    "S101", # asserts allowed in tests...
    "D100", # https://google.github.io/styleguide/pyguide.html#3821-test-modules
    "INP001", # https://docs.pytest.org/en/stable/explanation/goodpractices.html#which-import-mode
]
[tool.ruff.lint.flake8-annotations]
mypy-init-return = true
[tool.ruff.lint.isort]
force-single-line = true
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--strict-markers --cov={{cookiecutter.project_slug}} --cov-branch --junit-xml xunit-reports/xunit-result-ut.xml --cov-report xml:coverage-reports/coverage-ut.xml"
testpaths = ["tests"]
junit_family="legacy"

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/existing_code.html#introduce-stricter-options
strict = true
[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true

[tool.pyright]
include = ["src/", "tests/"]
exclude = [
    "**/__pycache__",
]
pythonVersion = "{{cookiecutter.minimum_python_version}}"
pythonPlatform = "Linux"
typeCheckingMode = "standard"
